(feat: 게시판 라우팅)

+++ src/components/board/BoardDetail.tsx
@@ -22,18 +22,23 @@ function BoardDetail(): JSX.Element {
 
   const handleDeleteSubmit = async (e) => {
     e.preventDefault();
-    const token = Cookies.get('key');
-    const headers = {
-      'Content-Type': 'application/json',
-      Authorization: `Bearer ${token}`,
-    };
-
-    const response = await api.delete(`http://localhost:7777/api/board/delete/${boardId}`, { headers });
-    if (response.data.status === 200) {
-      alert('게시글이 삭제되었습니다.');
-    } else {
-      alert(`${response.data.message}`);
-      navigate(-1); // 이전 페이지로 이동 (수정 폼 들어갔다가 나와서 목록으로 나가려는 사람들에 대한 예외처리 필요ㄴ)
+    const shouldDelete = window.confirm('정말로 삭제하시겠습니까?'); 
+  
+    if (shouldDelete) {
+      const token = Cookies.get('key');
+      const headers = {
+        'Content-Type': 'application/json',
+        Authorization: `Bearer ${token}`,
+      };
+  
+      const response = await api.delete(`http://localhost:7777/api/board/delete/${boardId}`, { headers });
+      if (response.data.status === 200) {
+        alert('게시글이 삭제되었습니다.');
+        
+      } else {
+        alert(`${response.data.message}`);
+      }
+      navigate(-2);
     }
   };
 
@@ -59,7 +64,10 @@ function BoardDetail(): JSX.Element {
             <TableBody>
               <TableRow>
                 <TableCell>
-                  <Typography sx={{display: 'flex', justifyContent: 'center'}}>[{(() => {
+                  <Typography sx={{ padding:1, color: 'grey', display: 'flex', justifyContent: 'left' }}> 
+                  {moment(board.moddate).format('YYYY-MM-DD HH:mm')}
+                  </Typography>
+                  <Typography sx={{display: 'flex', justifyContent: 'left', padding:1}}>[{(() => {
                 switch (board.category) {
                   case 'walk-with':
                     return '산책가요';
@@ -70,15 +78,20 @@ function BoardDetail(): JSX.Element {
                   default:
                     return board.category;
                 }
-              })()}] {board.title}   
-                    <Typography sx={{ color: 'grey' }}> {moment(board.moddate).format('YYYY-MM-DD HH:mm')}
-                    </Typography>
+              })()}] {board.title}  
                   </Typography>
-                  <Typography sx={{display: 'flex', justifyContent: 'center'}}> 작성자: {board.writerNickname}</Typography>
+                  <Typography sx={{display: 'flex', justifyContent: 'left', padding:1}}>
+                    👤 {board.writerNickname}
+                  </Typography>
                   <Divider sx={{margin:3}}/>
                   {board.image !== '' && <div style={{ display: 'flex', justifyContent: 'center', margin: '10px 0' }}>
                     <img src={board.image} alt={`${board.title}`} style={{ maxWidth: '100%' }} /></div>}
-                  <Typography sx={{display: 'flex', justifyContent: 'center', margin:5}}>{board.content}</Typography>
+                  <Typography sx={{display: 'flex', justifyContent: 'center', margin:5}}>
+                    {board.content}
+                  </Typography>
+                  <Typography sx={{display: 'flex', justifyContent: 'right', padding:1}}>
+                  👀 {board.clickCnt}
+                  </Typography>
                 </TableCell>
               </TableRow>
             </TableBody>
